Refactor sync.rs into submodules
Single 1050-line file was hard to navigate. Split into core.rs
(trait, shared types), push.rs, pull.rs, status.rs, and
mock_client.rs. No behavior changes — pure file reorganization.
