+++ crates/cloudsync-client/src/sync/mod.rs
@@ -1,0 +1,12 @@
+mod core;
+mod pull;
+mod push;
+mod status;
+
+pub use core::{DownloadFileResponse, SyncApi, SyncRecord};
+pub use pull::pull;
+pub use push::{CHUNK_SIZE, push};
+pub use status::status;
+
+#[cfg(test)]
+mod mock_client;
